Major Refactoring of code
Moved plenty of functions around
