+++ regression.py
@@ -15,15 +15,6 @@ with warnings.catch_warnings():
     import statsmodels.api as sm
 
 
-def list_recursive(d, key):
-    for k, v in d.items():
-        if isinstance(v, dict):
-            for found in list_recursive(v, key):
-                yield found
-        if k == key:
-            yield v
-
-
 def one_shot_regression(X, y, lamb):
     """Performs ridge regression
