+++ rw_utils.py
@@ -10,9 +10,7 @@ import os
 import ujson as json
 
 
-def write_file(spec, contents=None, location=None, file_name=None):
-    """Write contents to a file
-    """
+def return_file(spec, location, file_name):
     cache_dir = spec["state"]["cacheDirectory"]
     input_dir = spec["state"]["baseDirectory"]
     output_dir = spec["state"]["transferDirectory"]
@@ -20,6 +18,14 @@ def write_file(spec, contents=None, location=None, file_name=None):
     loc_dict = {"cache": cache_dir, "input": input_dir, "output": output_dir}
 
     args_file = os.path.join(loc_dict[location], file_name)
+
+    return args_file
+
+
+def write_file(spec, contents=None, location=None, file_name=None):
+    """Write contents to a file
+    """
+    args_file = return_file(spec, location, file_name)
 
     if contents is None:
         contents = spec
@@ -31,13 +37,7 @@ def write_file(spec, contents=None, location=None, file_name=None):
 def read_file(spec, location, file_name):
     """Read contents from a file
     """
-    cache_dir = spec["state"]["cacheDirectory"]
-    input_dir = spec["state"]["baseDirectory"]
-    output_dir = spec["state"]["transferDirectory"]
-
-    loc_dict = {"cache": cache_dir, "input": input_dir, "output": output_dir}
-
-    args_file = os.path.join(loc_dict[location], file_name)
+    args_file = return_file(spec, location, file_name)
 
     with open(args_file, 'r') as file_h:
         output = json.load(file_h)
